+++ consensus/src/marshal/mod.rs
@@ -56,6 +56,7 @@
 //!   uncertified blocks from the network.
 
 pub mod actor;
+mod request;
 pub use actor::Actor;
 pub mod config;
 pub use config::Config;
